requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
crewai==0.11.0
python-dotenv==1.0.0
langchain==0.0.350
openai==1.3.5
loguru==0.7.2 